requirements.txt
argparse
h5py
PyYAML
Pillow
numpy
scipy
torch
torchvision
pytorch-ignite
# tensorflow-gpu==2.0.0
tensorboardX==1.9

pytest
black
flake8